+++ xsolla-login-sdk/src/main/java/com/xsolla/android/login/callback/StartSocialCallback.kt
@@ -4,3 +4,8 @@ interface StartSocialCallback {
     fun onAuthStarted()
     fun onError(throwable: Throwable?, errorMessage: String?)
 }
+
+interface StartSocialLinkingCallback {
+    fun onLinkingStarted()
+    fun onError(throwable: Throwable?, errorMessage: String?)
+}