+++ xsolla-login-sdk/src/main/java/com/xsolla/android/login/entity/response/UrlToLinkSocialNetworkResponse.kt
@@ -1,0 +1,5 @@
+package com.xsolla.android.login.entity.response
+
+data class UrlToLinkSocialNetworkResponse(
+    val url: String
+)